@@ -4,9 +4,11 @@
   "description": "A twitter-like feed of all the channels you follow.",
   "main": "main.js",
   "scripts": {
-    "test": "npm run eslint",
+    "test": "npm run lint",
     "build": "webpack --progress",
-    "lint": "eslint src/"
+    "lint:js": "eslint src/ --ext js,jsx,json",
+    "lint:css": "stylelint src/*.css",
+    "lint": "npm run lint:js && npm run lint:css"
   },
   "repository": {
     "type": "git",
@@ -21,30 +23,58 @@
   "devDependencies": {
     "babel-core": "^6.21.0",
     "babel-loader": "^7.0.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-es2016": "^6.16.0",
-    "babel-preset-es2017": "^6.16.0",
+    "babel-preset-env": "^1.6.0",
     "babel-preset-react": "^6.11.1",
     "clean-webpack-plugin": "^0.1.10",
     "css-loader": "^0.28.0",
-    "eslint": "^4.2.0",
+    "eslint": "^4.7.2",
+    "eslint-import-resolver-webpack": "^0.8.3",
+    "eslint-plugin-compat": "^1.0.4",
     "eslint-plugin-freaktechnik": "^3.0.0",
+    "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-jsx": "0.0.2",
+    "eslint-plugin-react": "^7.4.0",
     "exports-loader": "^0.6.3",
     "extract-text-webpack-plugin": "^3.0.0",
     "html-webpack-plugin": "^2.22.0",
     "imports-loader": "^0.7.0",
     "json-loader": "^0.5.4",
     "style-loader": "^0.18.0",
+    "stylelint": "^8.1.1",
+    "stylelint-config-standard": "^17.0.0",
+    "stylelint-no-unsupported-browser-features": "^1.0.0",
     "webpack": "^3.0.0"
   },
   "dependencies": {
+    "babel-polyfill": "^6.26.0",
     "moment": "^2.14.1",
     "mprogress": "^0.1.1",
     "muicss": "^0.9.8",
+    "prop-types": "^15.6.0",
     "react": "^16.0.0",
     "react-autolink": "^0.2.1",
     "react-dom": "^16.0.0",
     "whatwg-fetch": "^2.0.2"
-  }
+  },
+  "stylelint": {
+    "extends": "stylelint-config-standard",
+    "plugins": [
+      "stylelint-no-unsupported-browser-features"
+    ],
+    "rules": {
+      "indentation": 4,
+      "color-named": "always-where-possible",
+      "plugin/no-unsupported-browser-features": [
+        true,
+        {
+          "severity": "warn"
+        }
+      ]
+    }
+  },
+  "browserslist": [
+    "last 2 versions",
+    "Firefox ESR",
+    "not ie <= 11"
+  ]
 }
