@@ -1,13 +1,16 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+
+const MENTION = "@";
 
 const Channel = (props) => {
-    var author = props.author || props.slug;
-    var tooltipAuthor = author.startsWith("@") ? author.substr(1) : author;
+    const author = props.author || props.slug;
+    const tooltipAuthor = author.startsWith(MENTION) ? author.substr(MENTION.length) : author;
 
-    return ( <a href={ "https://twitch.tv/" + props.slug } title={ tooltipAuthor + " on Twitch" }>{ author }</a> );
+    return ( <a href={ `https://twitch.tv/${props.slug}` } title={ `${tooltipAuthor} on Twitch` }>{ author }</a> );
 };
 Channel.propTypes = {
-    slug: React.PropTypes.string.isRequired,
-    author: React.PropTypes.string
+    slug: PropTypes.string.isRequired,
+    author: PropTypes.string
 };
 export default Channel;
