@@ -1,15 +1,18 @@
 import React from 'react';
 import Channel from './channel.jsx';
+import PropTypes from 'prop-types';
+
+const MENTION = "@";
 
 const Mention = (props) => {
-    var slug = props.text.toLowerCase();
-    if(slug.startsWith("@")) {
-        slug = slug.substr(1);
+    let slug = props.text.toLowerCase();
+    if(slug.startsWith(MENTION)) {
+        slug = slug.substr(MENTION.length);
     }
 
     return ( <Channel slug={ slug } author={ props.text }/> );
 };
 Mention.propTypes = {
-    text: React.PropTypes.string.isRequired
+    text: PropTypes.string.isRequired
 };
 export default Mention;
