@@ -1,6 +1,7 @@
 import React from 'react';
 import Panel from 'muicss/lib/react/panel';
 import Divider from 'muicss/lib/react/divider';
+import PropTypes from 'prop-types';
 
 import Avatar from './avatar.jsx';
 import Author from './author.jsx';
@@ -15,7 +16,7 @@ const Message = (props) => (
         <header>
             <Avatar height={70} width={70} className="mui--pull-left message-img" src={ props.avatar || defaultAvatar } />
             <Author slug={ props.authorName } author={ props.author } />
-            <a href={ "https://twitch.tv/" + props.authorName + "/p/" + props.id } rel="noopener" target="_blank" className="mui--pull-right mui--text-dark-secondary"><Timestamp date={ props.date } /></a>
+            <a href={ `https://twitch.tv/${props.authorName}/p/${props.id}` } rel="noopener" target="_blank" className="mui--pull-right mui--text-dark-secondary"><Timestamp date={ props.date } /></a>
         </header>
         <div className="message-container">
             <Divider/>
@@ -25,12 +26,13 @@ const Message = (props) => (
     </Panel>
 );
 Message.propTypes = {
-    avatar: React.PropTypes.string,
-    authorName: React.PropTypes.string.isRequired,
-    author: React.PropTypes.string.isRequired,
-    date: React.PropTypes.number.isRequired,
-    reactions: React.PropTypes.arrayOf(ReactionShape).isRequired,
-    id: React.PropTypes.node.isRequired
+    avatar: PropTypes.string,
+    authorName: PropTypes.string.isRequired,
+    author: PropTypes.string.isRequired,
+    date: PropTypes.number.isRequired,
+    reactions: PropTypes.arrayOf(ReactionShape).isRequired,
+    id: PropTypes.node.isRequired,
+    children: PropTypes.element
 };
 
 export default Message;
