@@ -1,6 +1,8 @@
 import React from 'react';
 import Mention from './mention.jsx';
 
+const NO_RESULT = -1;
+const START = 0;
 /**
  * This converts @name into a link to name's channel. However usually those are
  * mentions of the Twitter usernames, which sometimes don't match up. In a perfect
@@ -16,16 +18,16 @@ export default (nodes) => {
     const ret = [];
 
     for(const n in nodes) {
-        if(typeof nodes[n] == "string" && nodes[n].search(/@\S+/) != -1) {
+        if(typeof nodes[n] == "string" && nodes[n].search(/@\S+/) != NO_RESULT) {
             let node = nodes[n];
             let result = node.search(/@\S+/);
             let offset = 0;
-            while(result != -1 && node.length) {
-                if(result > 0) {
-                    ret.push(node.slice(0, result));
+            while(result != NO_RESULT && node.length) {
+                if(result > START) {
+                    ret.push(node.slice(START, result));
                 }
-                const mention = node.match(/@\S+/)[0];
-                ret.push(<Mention text={ mention } key={ n + ":" + (offset + result) + mention }/>);
+                const mention = node.match(/@\S+/).shift();
+                ret.push(<Mention text={ mention } key={ `${n}:${offset + result}${mention}` }/>);
 
                 if(result + mention.length < node.length) {
                     node = node.slice(result + mention.length);
@@ -34,7 +36,7 @@ export default (nodes) => {
                 }
                 else {
                     node = "";
-                    result = -1;
+                    result = NO_RESULT;
                 }
             }
             if(node.length) {
