@@ -6,7 +6,7 @@ import "babel-polyfill";
 
 import Page from './page.jsx';
 
-import 'muicss/lib/css/mui.min.css';
+import 'muicss/dist/css/mui.min.css';
 import 'mprogress/build/css/mprogress.min.css';
 import './style.css';
 
