@@ -8,22 +8,24 @@ const opts = Object.freeze({
 });
 
 const getUserID = async (username) => {
-    const response = await fetch(`https://api.twitch.tv/kraken/users?login=${username}`, opts);
+    const response = await fetch(`https://api.twitch.tv/kraken/users?login=${username}`, opts); // eslint-disable-line compat/compat
     if(!response.ok) {
-        throw `Could not load user ID for ${username}`;
+        throw new Error(`Could not load user ID for ${username}`);
     }
     const json = await response.json();
-    if(json.users.length < 1) {
-        throw `No users found for username ${username}`;
+    if(!json.users.length) {
+        throw new Error(`No users found for username ${username}`);
     }
-    return json.users[0]._id;
+    return json.users.shift()._id;
 };
 
-const getFollows = async (userId, page = 1) => {
-    const response = await fetch(`https://api.twitch.tv/kraken/users/${userId}/follows/channels?limit=${followsLimit}&offset=${(page - 1) * followsLimit}`, opts);
+const FIRST_PAGE = 1;
+
+const getFollows = async (userId, page = FIRST_PAGE) => {
+    const response = await fetch(`https://api.twitch.tv/kraken/users/${userId}/follows/channels?limit=${followsLimit}&offset=${(page - FIRST_PAGE) * followsLimit}`, opts); // eslint-disable-line compat/compat
     const json = await response.json();
     if(json._total > page * followsLimit) {
-        const follows = await getFollows(userId, page + 1);
+        const follows = await getFollows(userId, page + FIRST_PAGE);
         return json.follows.concat(follows);
     }
     return json.follows;
@@ -32,7 +34,7 @@ const getFollows = async (userId, page = 1) => {
 const getPosts = async (userId) => {
     let json;
     try {
-        const response = await fetch(`https://api.twitch.tv/kraken/feed/${userId}/posts`, opts);
+        const response = await fetch(`https://api.twitch.tv/kraken/feed/${userId}/posts`, opts); // eslint-disable-line compat/compat
         json = await response.json();
     }
     catch(e) {
